fixed OpenCL n_body with float

+++ test/ocl/n_body/n_body.cl
@@ -19,8 +19,8 @@ __kernel void n_body(__global body* B_read, __global body* B_write, int size) {
 	for(int k=0; k<size; k++) {
 		body bk = B_read[k];
 		triple dist = SUB(bk.pos, bgid.pos);
-                double r = ABS(dist);
-                double f = (gid == k) ? 0 : (B_read[gid].m * B_read[k].m) / (r*r);
+                float r = ABS(dist);
+                float f = (gid == k) ? 0 : (B_read[gid].m * B_read[k].m) / (r*r);
                 force cur = MULS(NORM(dist), f);
                 F = ADD(F, cur);
 	}
